requirements.txt
flask
flask_sqlalchemy
python-docx
gunicorn
pymysql
python-dotenv
psycopg2-binary
